requirements.txt
openstacksdk>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
requests>=2.25.0 